Move board component to appropriate folder
